BUGFIX
the snort "msg" section does not handle parens [()] or semi-colons [;]
well AT ALL. as in, the rule fails to validate if these symbols are
there (even though they're encased in double quotes).
I implemented three ".replace" functions to strip the open paren [(],
close paren [)] and semi-colon [;] out of the message portion of the
rule.